rename contact to user, move account to user, add serialize test
